Better UX experience. remove members UI side and fix according to request result.

+++ src/components/GroupMembersList/groupActions.ts
@@ -5,8 +5,8 @@ import {
 import { Api } from '../../services/api'
 import { showErrorMessage } from '../App/Snackbar/snackbarActions'
 import { selectToken } from '../App/userSelector'
-import { selectGroups } from '../../pages/AdminDashboard/groupsSelector'
-
+import { selectGroup } from './groupSelector'
+import _ from 'lodash'
 
 
 export const setGroup = (group: any): GroupActionsTypes => ({
@@ -30,12 +30,25 @@ export const getGroup = (t, groupId) => async (dispatch, getState) => {
 
 export const deleteMembers = (t, groupId, members) => async (dispatch, getState) => {
     const token = selectToken(getState())
+    const group = selectGroup(getState())
     const api = new Api(token)
+
+    if(!group?.members){
+        return
+    }
+
+    const newGroup = {
+        ...group,
+        members :  _.differenceBy(group.members, members,'id')
+    }
+    
+    dispatch(setGroup(newGroup))
 
     try {
         await api.deleteMembers(groupId, members)
         dispatch(getGroup(t, groupId))
     } catch (err) {
+        dispatch(setGroup(group))
         dispatch(showErrorMessage(t('errors.delete-members')))
     }
 }
